[refactor] extract express entrypoint
